Update patch set 5
Patch Set 5:

(1 comment)

After thinking a bit more, maybe DndSdInstanceRecord and DnsSdInstanceEndpoint aren't the best names. The idea is:

DnsSdInstanceRecord = A specific service which is hosted by a machine. Ie, it could be "Chromecast is running at this port" or "spotify is running at this port" or "XXXXX is running at this port"

DnsSdInstanceEndpoint = The network interface endpoint associated with a given service running on a machine. It's closer to "Chromecast is running at this port + IP Address, so use this endpoint to talk to it"

This means "XXXXX is running at this (port + IP)" is an "XXXXX is running at this port", where there might be 1 or more IPs associated with the given service (one per network interface)


Naming ideas (instead of DnsSdInstanceEndpoint is a DnsSdInstanceRecord):
DnsSdServiceInstance is-a DnsSdServiceType
DnsSdInstanceRecord is-a DnsSdInstance
DnsSdDiscoveredEndpoint is-a DnsSdDiscoverableService

Patch-set: 5
Reviewer: Gerrit User 1330244 <1330244@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Commit-Queue=0