Cast Streaming Standalone Sender: Congestion control, and video.
Add congestion control logic to the standalone sender app, updating the
target encode bitrate of the audio and video encoders based on estimated
network bandwidth availability. Added a --max-bitrate command-line
argument  the default max bitrate is 5 Mbps).

Plumbs-through video to the StreamingVp8Encoder, such that the
standalone sender is now complete with both audio and video!

Change-Id: Icad946ab91be1f7689c0b02852f7c3e462030393