Add ARM64 build bot
This patch adds a bot that builds on linux64 for a target cpu of ARM64.
After this patch lands, we should have validation as part of checkin
that our cross compile toolchain is working properly.

Next step will be updating the scripts/slave/recipes/openscreen.py file
to run the tests (QEMU or isolate TBD).

Bug: openscreen/18
Change-Id: I744403788d055e09ab27931c8141dba09b268c58