Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> Generally a factory method (TryCreate) should replace a constructor.
> After this CL in mdns_records some classes have both TryCreate and a constructor, and some have a constructor only.
> Ideally, there should be one way only to construct a class, as this eliminates ambiguity.
> Since you are guarding against the data received over a network, I think the validations in TryCreate belong in mdns_reader instead, where other validations already take place.

I am open to changing everywhere to use the TryCreate() methods, and adding them in the remaining classes that don't have them. That definitely sounds less ambiguous

The reason I put the logic here instead of in mdns_reader is that the latter would involve duplicating a lot of logic (a lot of the DCHECK'd values in mdns_records) into mdns_reader for no added benefit. IMHO it seems safer to avoid all DHECKs in mdns_reader - not to mention more readable since the same validations are used in all places that each type of Record/Rdata/DomainName/etc.. is used

WDYT?

Patch-set: 2