Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> > Patch Set 2:
> > 
> > Generally a factory method (TryCreate) should replace a constructor.
> > After this CL in mdns_records some classes have both TryCreate and a constructor, and some have a constructor only.
> > Ideally, there should be one way only to construct a class, as this eliminates ambiguity.
> > Since you are guarding against the data received over a network, I think the validations in TryCreate belong in mdns_reader instead, where other validations already take place.
> 
> I am open to changing everywhere to use the TryCreate() methods, and adding them in the remaining classes that don't have them. That definitely sounds less ambiguous
> 
> The reason I put the logic here instead of in mdns_reader is that the latter would involve duplicating a lot of logic (a lot of the DCHECK'd values in mdns_records) into mdns_reader for no added benefit. IMHO it seems safer to avoid all DHECKs in mdns_reader - not to mention more readable since the same validations are used in all places that each type of Record/Rdata/DomainName/etc.. is used
> 
> WDYT?

Main reason why constructors have DCHECKs is that all the code in the library should be developed in such a way that when DomainName, any Rdata, MdnsRecord, etc. are created in the code they are always created as a valid instance of a type, and we know that there's no need to check a result of TryCreate and that there are no exceptions happening either (or DCHECKs catch that in debug builds when we run tests). This way we only need to perform check when parsing untrustworthy input (in this case - received over the network).

This approach was picked because it results in 1) cleaner call sites throughout the code and overall simplifies the code that works with these classes, 2) fewer unnecessary runtime checks for things that would never happen. The cost, of course, is that validation check is in the mdns_reader and inside the constructor in a DCHECK.

I've made a clear choice towards the approach without a TryCreate when I wrote the code.

Patch-set: 2
Reviewer: Gerrit User 1336186 <1336186@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>