Fix gcc bot
This change makes some -Wno-* flags clang-only and adds some that are
gcc-only because the gcc bot suddenly changed its mind about the flags
it recognizes and the errors it emits.  It also fixes signed integer
wrapping undefined behavior in a bandwidth estimator test, which also
seems to have been exposed by the gcc change.

Change-Id: I777c9bf2e7e7f08f5f6ff479bbd927f378dc63a6