Upgrade ARM64 bot configuration
This patch updates the ARM64 bot from Linux 1604 to Linux 1804, in order
to request the new docker image for swarming.

Note: a follow up patch in the openscreen.py infra file is likely.

TBR so that I can test bot stability over the weekend, and it's a
minimal change that shouldn't require much review.

TBR=btolsch@chromium.org
Bug: openscreen/18
Change-Id: I27d6728dee49375c0441d113277c3a3e13b8d44e