Discovery: Update contract between discovery layers
This CL updates documentation and source code to fit the new method
contract that callbacks from lower layers will not invoke calls back
to that same layer

It addresses comments in the following CL while providing an alternative
solution to the bug:
https://chromium-review.googlesource.com/c/openscreen/+/2080670

This fix was added due to the following edge case:
The OnRecordChanged callback would fire kExpired for a PTR record. This
would make dnssd/impl/querier_impl stop the mDNS kALL query associated
with the pointed to domain. This would then call mDNS Querier and update
|callbacks_| which would invalidate the iterator still being used by the
original call, resulting in an infinite loop since the exit condition is
never hit.

Additionally, E2E tests have been updated to test for this edge case.

Change-Id: Ie1a8301cfb5e8589a83b3015571fac79edca15af