Code Coverage: Unchanged Chromium Scripts
This CL copies 2 scripts unchanged from Chromium, to be used by the
buildbots.

In chromium, these scripts are run during the build, after the shareded
tests generate '.profraw' files for code coverage data, in order to
gather them up, merge them, and convert them to a usable '.profdata'
file.

Here, the merge step is unneeded, but the conversion is still needed.
This will be run by our buildbot, after the code_coverage module has
been initialized and unit tests have been run, but before the generated
coverage data can be used and uploaded.

NOTE: These files are entirely unchanged from Chromium. It's possible
there are issues, but my preference is to check them in unchanged so
updates made in Chromium can be rolled with minimum effort, on an as-
needed basis.
Change-Id: I672e8a60ea2437fb5eb3eaa3f8c37d00a0810879