Code Coverage: Unchanged Chromium Scripts
This CL copies 2 scripts unchanged from Chromium, to be used by the
buildbots.

These 2 scripts are used as follows:
merge_steps.py:
    This script is responsible for merging multiple '.profdata' (code
    coverage data files) into a single shared file. In chromium, this
    is useful for merging multiple sharded tests' profdata files back
    into a single merged file on the root machine.

    In openscreen, because all tests are run on the same machine, this
    script is not helpful. That being said, this script is called from
    the chromium code_coverage module. Its path must be provided to
    this script, or code coverage run on the buildbot will fail.

merge_lib.py
    This file provides the functionality backing merge_steps.py.
    Additionally, it is used by the merge_results.py script in the
    follow-up CL.

NOTE: These files are entirely unchanged from Chromium. It's possible
there are issues, but my preference is to check them in unchanged so
updates made in Chromium can be rolled with minimum effort, on an as-
needed basis.
Change-Id: I672e8a60ea2437fb5eb3eaa3f8c37d00a0810879