Fix CastSocket e2e tests
This patch fixes the Mac and ARM64 cast socket end to end tests. Turns
out this was actually a bug in the SocketHandleWaiter, where it
would only ever process the first ready handle in its list of handles,
completely starving any other handles in its list.

The fix is to store the last updated time along with the socket
subscriber, and continously process the last updated handles until
either (1) a timeout is reached, or (2) all handles have been updated.

Bug=b/155214294
Change-Id: Icf967a80036654520a3d13661a596f38730951c4