Move Code Coverage to new FYI bot
Currently, the Step module's step.nest command as used in the openscreen
trybot and the code_coverage module does not function as documented. This
is a known issue and is being tracked by bug:
https://bugs.chromium.org/p/chromium/issues/detail?id=1078530

Until this bug is resolved, code coverage must be calculated in a
separate trybot, or issues in code coverage can cause an otherwise
passing trybot to fail.

Change-Id: I47a9bf8429ce424262660b13911e1f14d14f49e4