Move Code Coverage to new FYI bot
Currently, the Step module's step.nest command as used in the openscreen
trybot and the code_coverage module does not function as documented:
https://cs.corp.google.com/github/luci/recipes-py/recipe_modules/step/api.py?fromv2=true&l=208

This is a known issue and is being tracked by bug:
https://bugs.chromium.org/p/chromium/issues/detail?id=1078530

Until this bug is resolved, code coverage must be calculated in a
separate trybot, or issues in code coverage can cause an otherwise
passing trybot to fail. This CL adds an FYI bot to calculate code
coverage, and removed the code coverage mixin from existing required
bots.

Change-Id: I47a9bf8429ce424262660b13911e1f14d14f49e4