Update patch set 2
Patch Set 2: Commit-Queue+2

Here's what I could find out based on some looking around.  The C++ lint checker is written in Python and based on regular expressions, so it has some false positives.  It seems to have trouble with multi-line constructors.

There seem to be three forked copies of the linter and the one in Chromium has diverged from the two in GitHub.  It's not clear how much maintenance the Chromium one is getting these days.  

We can probably turn off most of the syntax checks in the future as they should all be handled by clang-format.  But not quite all, as things like the copyright, #include format, ending newlines etc. aren't going to be handled by a source parser.

Short version: if a linter check is not providing value, it's easy to turn it off.

Patch-set: 2
Reviewer: Gerrit User 1001982 <1001982@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Commit-Queue=+2