Code Coverage: Add new buildbot mixin
This CL adds a new buildbot mixin that denotes whether a given build is
for a 'ci' but or a 'try' bot. This is needed because the same code
coverage module is used for CL code coverage and for full-repo code
coverage, with an input flag to the module differentiating the two.

Change-Id: I3b3a69a56f3b3e2b98ad718e44c91d5eecc49ffe