Update patch set 5
Patch Set 5:

(16 comments)

1) The procedural coding pattern for this parser is quite repetitive.  There's one pass to do syntactic parsing for each field, then another pass to do semantic validation, resulting in at least two if/else clauses for every field - and there are a *ton* of fields.

A common pattern to write a parser is to have an object that visits each token in the parsed input, does syntactic and semantic validation, and fills in a target structure with values.  The parser is written in a data driven style where it consults a template to know what kind of value to expect for each input token.  A parsing error throws an exception which is routed into one exception handler.

What makes this annoying in Chromium C++:
- Doing the data driven parsing thing requires a lot of template programming.
- Handling errors is a pain without exceptions.

I think it's possible to do it this way, but I wonder if it is worth it or not, given that all of this code has already been written.  I guess you could invent some macros to reduce the boilerplate, but that just papers over the problem.

At the end of the day, this protocol is not going to change in significant ways, this implementation will catch all the errors, and it's good enough.

2) I'm wondering what a good test suite for this parser is.  You'll need a lot of test cases to cover all the possible errors. Can you give me some insight into what are the most important errors to catch?  Are there enough test cases for them?

Given the number of possible error cases, I would suggest thinking about testing against a file of bad JSON answer messages rather than having to write code for each one.

Patch-set: 5
Reviewer: Gerrit User 1001982 <1001982@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>