Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> (3 comments)
> 
> The opt records should be a DnsType like others, that would make this cleaner.
>
> It looks like Android is utilizing option code 4 for mDNS sleep proxy support.  We want to do this eventually for OSP, so this isn't just an interop fix.
> 
> https://www.ietf.org/archive/id/draft-cheshire-edns0-owner-option-01.txt

The approach here was intended to localize all changes to as-few-files-as-possible - this allowed us to parse the record as expected, but then treat it as an unknown record type for the rest of the code. Adding a new record type adds a lot of edge cases elsewhere and I wanted to avoid that.

If we want to support OPT pseudo-records long term, I'm going to put this review on hold and examine the approach. I want to do some research into the other pseudorecord types, as well as allow for more detailed MdnsRecord child class which parses into a full record type. This may warrant a larger design, depending how the investigation goes - I expect it will add a significant amount of complexity

If this does end up a larger task, WDYT about committing this or similar as a short-term fix (specifically to address the attached bug), then reverting it when full support can be added?

Patch-set: 3
Reviewer: Gerrit User 1330244 <1330244@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Commit-Queue=0