Update patch set 20
Patch Set 20: Code-Review+1

(9 comments)

Looks great, thanks for sticking with this. 

Regarding platform/base/{optional,simple_fraction} this is not a thing I'm super happy with, as it's a consequence of exposing the offer/answer data structures in our public API.  I wasn't sure if this was a deliberate design decision for the streaming module or a matter of convenience, but either way it means clients can become dependent on the exact structure of the answer which will make it harder in the long run to iterate on the protocol.

Also it exposes our implementation of Optional which is different from std::, absl:: and base::, and won't be convenient for clients to use (they will need to convert).

But this shouldn't block landing this, if we are able to hide these details soon the risk of clients adding dependencies is low.

Patch-set: 20
Reviewer: Gerrit User 1001982 <1001982@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Code-Review=+1