Add unit tests for IPAddress, IPEndpoint comparisons
The comparison operators for IPAddress, IPEndpoint are somewhat
unintuitive--what does it mean for an IPAddress to be greater or less
than a different one?

This patch improves clarity around this question by adding comments and
unit tests, with the added bonus of improved test coverage.

Adding these tests showed that some of the operators were broken/buggy, so this
includes some fixes.

Change-Id: I49bf770a14d8b4ed804004d59a5800583a3b1a31