Discovery: Update Public API to support multiple addresses
Currently, the DNS-SD implementation only supports a single A and AAAA
record per received service instance. This works for CastV2, but as a
more general mDNS + DNS-SD implementation, this limitation should be
removed. This CL takes the first step towards this change by updating
the public API to support a vector of addresses rather than a single
address.

Bug: b/158531165
Change-Id: I0c289428be1e558b0b7a7dede86dbe18aefafd75