Discovery: Update DNS Querier
This CL implements the swap from dns_data to dns_data_graph, as added
in CL:
https://chromium-review.googlesource.com/c/openscreen/+/2223690

After the completion of this CL, support for Address records with domain
different from that of the SRV record pointing to them will be supported.
See documentation at the top of dns_data_graph.h for further details
about this hierarchy.

NOTE: After this change, the querier will maintain support for multiple
address and SRV records as called out in b/158533685 and b/158532090.

Bug: b/157683753
Change-Id: Ibb5b0786149cdb81ae11a551626f77ac15117dc2