Discovery: Add support for subtypes to the public headers
This CL adds support for Subtypes in the DnsSdInstance object, as
defined in https://tools.ietf.org/html/rfc6763#section-7.1
In future, as part of b/158534054 ,support for subtypes / Selective
Instance Enumeration will be fully added. For now, just the public
APIs are being updated to prevent breaking users when these APIs are
later implemented

Here, DnsSdInstance and DnsSdInstanceEndpoint are publicly exposed
classes. It is expected both that embedders will directly be interacting
with these classes and, if they choose to override our DND-SD
implementation (As is expected in many cases), becoming intimately
familiar with them. For this reason and similicity of use by the
embedder, support for a variety of ctors has been added to both
DnsSdInstance and DnsSdInstanceEndpoint.

Bug: b/158533407
Change-Id: I31eb9433a8e4ecb6ef018327a937e33b7e827ab1