Cast Streaming: Provide a reason code when destroying Receivers.
This allows clients to determine whether they need to do a full tear-
down (at end of session), or just prepare for a new set of Receivers
(the result of a re-negotiation).

In addition, the OnConfiguredReceiversDestroyed() method was renamed to
OnReceiversDestroying() to correct the tense of the verb, and make the
name unambiguously shorter.

Bug: 1116185
Change-Id: Ib3a14834c47997608bd719450b9755f452fa3859