Add note about Abseil StrFormat
Currently, we don't have StrFormat in the code base due to binary size
concerns (it adds ~350kB to the openscreen_unittests binary, and
presumably will add similar amounts when our library is linked in
projects that don't use Abseil). This patch removes the str_format.h
header from our Abseil source set, as well as adds a comment to ensure
that this exclusion is continued in the future.

Change-Id: I1a7055d64f1e3ec995adc970cf6d74ea6850b1d4