Fixup standalone_{receiver, sender}
Currently the cast sender and receiver cannot connect because the
receiver cannot present a CA-signed certificate due to not being able to
set the private key or use a self signed certificate.

This patch adds support for setting these arguments on both the cast
sender and receiver, as well as some fixes found through manual testing.

Example usage:
$ ./out/Default/cast_sender -s cast_streaming.crt -v ~/video-1080-mp4.mp4
$ ./out/Default/cast_receiver lo0 -v -x -p cast_streaming_rsa -s cast_streaming.crt

Bug: b/156995806
Change-Id: I6e31e66beff33c260e467290f454ec1dcc758660