Add build flag for self-signed certs
This patch changes the standalone receiver and sender implementations to
use a new CAST_ENABLE_ALTERNATE_CERTIFICATE build flag/preprocessor
macro for gating usage of self-signed certificates. When false, the
cast_receiver app is disabled, and the cast_sender app will only connect
to receivers using a proper Google cert.

When the build flag is enabled, the cast sender can use any self-signed
certificate, for a maximum of 48 hours after its valid after time. The
cast receiver can either take a self-signed certificate and its private
key, generate a certificate from a passed in private key, or generate
both the private key and the certificate. The resulting certificate is
written out to a file (currently <build output
dir>/generated_root_cast_receiver.crt) and can then be used with a
cast_sender to start a session.

Bug: b/169796278
Change-Id: I03675f85ac0b2bda76daf1bf11d9d7df064d0b8f