LICENSE file checking
This patch adds license file checking in Open Screen, by adding a new
licenses.py that iterates through third_party dependencies and returns
a list of licensing errors. This runs as part of PRESUBMIT.

This patch also includes README.chromium and LICENSE fixes for
third_party dependencies.

Finally, as part of refactoring I fixed the DEPS checker--before this
patch it fails to load modules so never runs on pre-submission.

Bug: b/173625891

Change-Id: I417e61f878dab809cf959d69480be749f9229128