Fix supported_address_families DCHECK in ServiceInstance ctor.
The original DCHECK was mandating IPv4/6 be used if the network
interface supports it, even if a client may not want to use IPv4/6.

This patch changes the logic to the intended sanity-check: IPv4/6 can
only be considered supported if the network interface has an IPv4/6
address. The client can still choose whether to use it or not.

Bug: b/162542369
Change-Id: I76e3983eb36c63769d4574e75e07df51dd4540d4