[HOTFIX] Fix Mac standalone build
Currently the standalone build doesn't work on the bots, so it tends to
break without warning.

This patch fixes an issue where a static member variable that has a
definition but not a separate declaration causes a linking error on Mac
OS X. Not sure why it doesn't happen on other platforms.

For more information on this issue, see:

https://stackoverflow.com/questions/8016780/undefined-reference-to-static-constexpr-char

TBR=miu@chromium.org

Change-Id: I97a7624ef77f85ed0995052d6db5e28834a88dff