Full Cast Standalone Sender (Mirroring)
This patch adds all the necessary sender-side logic for talking to the
agent on a Cast Receiver, to LAUNCH the Mirroring App, and to estabilsh
a VirtualConnection for Mirroring App message routing.

It heavily modifies the role of LoopingFileCastAgent: This class now
becomes a "single-use workflow" that walks through all the steps needed
for establishing streaming and tearing it down.

With this patch, the Standalone Sender can now be used with the
Standalone Receiver once again.

Bug: b/162542369
Change-Id: Ib7d990995b5a036bd0299806a42d4a92f89d9209