Fix UDP bind timing issue
This patch fixes a potential timing issue where we attempt to send an
Answer message before the UDP socket is bound.

The current behavior is
that we construct an Answer with a UDP port value of 0, which results in
the Answer failing validation and an error Answer message being sent
back to the sender.

The new behavior is that the Answer message is not sent until the
environment is notified that the UDP socket has been bound. This is
accomplished by an addition to the UDPSocket::Client abstract class, a
new OnBound event.

Bug: b/174501332
Change-Id: I4bcebe78d8d41c29bb4b4698c5565f0dd9ef2c17