Use X509 getters instead of reaching into BoringSSL internals
This CL is needed to compile with the latest revision of BoringSSL.

We're in the process of aligning with OpenSSL upstream and switching
the X509 structure from direct struct access to accessors. This CL fixes
most instances leaves the mutation of x509->nc and x509->ex_flags
(which, even before the struct became opaque, was unsupported), since
they're a little more subtle.

This also switches from X509V3_EXT_d2i to X509_get_ext_d2i, which saves
some code. Note, however, both the old and new version of the CL do not
correctly handle encoding errors in the extensions. For this CL, I've
preserved the original bugs and have focused on resolving the direct
struct accesses, rather than doing a thorough review for correctness.

Change-Id: Ib7c535b726684719baa15c76a84e95a18e0d8114