Maintain self-issued bit in a local variable
EXFLAG_SI is recomputed in each loop anyway, so there is no point
storing it with the certificate in the first place. It is also, in this
code, not *entirely* path-independent due to the is_root check. This
fixes a potential bug where checking one path impacts the behavior of
another path.

Change-Id: If9b0d157dd49c44723c1a8e2f83eebfc3fc2779c