[Discovery] Fix bug in PTR expiry
This patch fixes a bug where PTR expire updates with a different TTL
from the current record are not considered to be updates, leaving the
now expired record in the graph and returning an error.

A new equality operator, "IsUpdateOf", is added to MdnsRecord to allow
for checking this specific case.

Bug: b/168800070
Change-Id: I5c5d4372d3ded791d102cb95cc37fdf5e944d404