[Cast] Fix Standalone Sender/Receiver
Over the last year, the standalone sender and receiver applications have
gone through some major changes, and unfortunately a few regressions
have been introduced.

This patch fixes the poor performance. This was actually caused by a "fix" to the
PlatformClientPosix implementation, where we were using inappropriate
values for network looping (50ms, which is an eternity).

Bug: b/182937147
Change-Id: I39af927401ce940b81829e5b2992e06bc544ef01