[Cast] Fix bandwidth estimator
Currently, the bandwidth estimator always reports 0kbps, even though we
are successfully sending packets.

Turns out, the LoopingFileSender is erroneously using a second
SenderPacketRouter that is separate from the SenderSession's instance,
so it has no packets to report. This feature is now exposed on the
SenderSession, since bandwidth estimation is a helpful signal for
embedders to throttle capture.

Bug: b/183996401
Change-Id: I932085dfe00df29d8ae79a184ee10b1d3fe5594d