Remove OperationLoop
This change removes OperationLoop because it's only used in one place
and is arguably confusing.  OperationLoop mixes sleep() and select()
usage, with separate timeouts, which has resulted in some "tuning" bugs
in the past.  Since there's no benefit to the sleep portion and it is
also what causes the confusion, this change removes it and thing
directly uses SocketHandleWaiter in PlatformClientPosix.

Bug: None
Change-Id: I9f62c9253bee7d29b26d1838d34c1399f4d472b5