[Cast] Cleanup Resolution + ReceiverSession
This patch changes the way Dimensions and Resolution objects are
used in Open Screen--consolidating them into just one set
of Dimensions and Resolution objects that can be used throughout
the library, and updates the mirroring control protocol specification to be
crisper about the difference between Resolution objects and Dimensions
objects--and removing frame rates where they don't really make sense.

As part of this cleanup, it also refactors the
ReceiverSession to not expose the Answer messaging system at all,
instead using new publicly exposed types. Finally, the capture
recommendations are removed from the Receiver side, in favor of shared
constants.

Bug: b/184186374
Change-Id: I5e6a1c6798d07cb7f388ff5179c0bfa325d89229