[Cast Streaming] Add codec negotiation over RPC
This patch adds support for negotiating codec information over the RPC
messenger class for both the standalone sender and receiver.

Although a full remoting implementation in the standalone
implementations is not a goal and not particularly practical, it is
important that we exercise the public remoting APIs and get basic usage
and flows handled in our library.

To that end, this patch specifically adds support for sending
RPC_DS_INITIALIZE and RPC_DS_INITIALIZE_CALLBACK messages, simulating
the process of setting up a remoting media stream and demuxer in Chrome.

NOTE: this does not enable "true" remoting--the standalone sender is
still decoding and reencoding the video. For "true" remoting support we
need the ability to unpack raw frames (probably exclusively from a WebM
container that holds VP8 frames) and pass them directly.

In future patches, we can also consider adding additional features to
the standalone implementation, such as play/pause and volume control
methods.

Bug: b/190078859
Change-Id: I49c2acb537aa3c647b8c3a53c9545f5eb3893982