[Open Screen] Align compiler warnings list with Chromium.
Chromium is preparing to enforce an expanded list of compiler warnings
(https://crbug.com/1213098).  This aligns the clang compiler warnings
with those that will be enforced in Chromium, fixes multiple
violations in our code, and allows some violations in third_party code.

The exception is -Wshorten64-to-32, which I'd like to consider
implementing our own checked_cast library functions for.

Notes:

- I am using unchecked casts from long to double() for some
  std::chrono::duration values.  No precision is lost for values
  under 2^53 and I don't believe casts these are problematic, but
  "double" check.
- I did check what happens with these flags on gcc.  If we want dual
  enforcement I'd like to deal with that in a followup change.

Bug: b/200962781
Change-Id: Ieb1830fd2ac74566e654be580c7b5c243975421f