Switch the recipes to use bbagent.
Generating the config using bbagent changes the proto message to use the
exe field instead of recipe, so the generated config is different, but
the builders are already running using bbagent because buildbucket is
globally forcing it on. So this change should effectively be a no-op. An
upcoming change to switch recipes to use python3 will also cause the
switch to the new proto message, so this change will isolate those
modifications to cr-buildbucket.cfg.

Change-Id: I29a61502bf265ba784e87041da9b119beff5ae23