Avoid tautological comparison to 0 when char is unsigned.
The `char` type can be either signed or unsigned. If it is unsigned,
comparing to see if it's >= 0 is always true, and generates a warning
with -Wtautological-unsigned-zero-compare.

This warning is useful for looking for bugs. To avoid this false
positive, fix the code to work without warnings when char is signed
or unsigned, by specifying the type explicitly as unsigned. Then, the
bits therein will always be interpretted as a positive number, and we
can compare to just `<= 9` instead of `>= 0 && <= 9`. After that
change, no warnings will be emitted regardless of unqualified-char's
signedness.

Bug: chromium:1287175
Change-Id: Iff5c570bbdbfa7b294a1762b94e68f88f4cb8912