[Open Screen] Roll abseil.
Abseil is now being rolled more aggressively in Chromium to near ToT.
Since our Linux build uses the Chromium toolchain which is also kept
near head, we need to disable new compiler warnings that are yet to be
fixed in abseil.

We don't need abseil ToT for anything specific so just use the most
recent long term stable ref.

On Mac, we use the XCode toolchain which does not have the new
warnings.

Change-Id: I7dbf4ad2c8c5e4f4b39a0b55ac05ed0ff9327e2f